hwc: Fix regression in hwc_sync()
hwc_sync() is located in a hotpath of a display rendering code, hence it
is undesirable to process any I/O operations within it. However, a property
"debug.egl.swapinterval" is getting read during each sequence of that call,
making much more latency than there should be.

Omit swapinterval usage if an appropriate flag is not specified in a device
tree, thus reducing the latency of hwc_sync() call and making HW Composer
process faster on older hardware with slower flash storage.

Test: Manual testing for proper conditioning.
Test: Running full CTS suite without DISPLAY_DEBUG_SWAPINTERVAL set.
Change-Id: If08595d37ac0c1a19edd8d296289a21c9f95512a
Signed-off-by: Alex Naidis <alex.naidis@linux.com>